Added functionality to achieve below acceptance criteria:
I see a list of the album's songs.
When I click on a song, it plays.
When I click on a playing song, it pauses.
When I hover over a song, it displays a "play" button in place of the song number.
The currently playing song displays a "pause" button in place of the song number.
A paused song displays a "play" button in place of the song number.