Added functionality for below items:
Regardless of scroll position, I see a "player bar" with controls for
Play/Pause, Previous Track, Next Track, Seek, and Volume.
The Play/Pause button shows a "play" icon if no song is playing.
The Play/Pause button shows a "pause" button if a song is playing.
The Next Track button plays the next song on the album.
The Next Track button doesn't respond if there is no next song.